requirements.txt
fastapi>=0.115.0
uvicorn==0.24.0
pydantic[email]>=2.9.0
pydantic-settings>=2.4.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
Jinja2==3.1.2
python-dotenv==1.0.0
SQLAlchemy>=2.0.36
pg8000==1.30.5
